api/requirements.txt
requests
fastapi
uvicorn
pandas
pyarrow
psycopg2-binary
python-dotenv
pytest